+++ mr_poopybutthole/snowflake.py
@@ -20,7 +20,7 @@ class Snowflake(commands.Cog):
         self.logger = logging.getLogger(__name__)
         self.bot = bot
         self.thelist = yaml.load(open(SNOWFLAKES_FILE), Loader=yaml.FullLoader)
-        self.snowflake_list = {} 
+        self.snowflake_list = {}
         for snowflake in self.thelist.keys():
             self.snowflake_list[snowflake] = True
         self.snowflake_mode = False
@@ -28,26 +28,30 @@ class Snowflake(commands.Cog):
     async def set_snowflake_name(self, snowflake, member, reset=False):
         """
         Set or reset the nickname of a specific snowflake. Requires the
-        snowflake name, their member object (which can beobtained from
-        guild.fetch_member()), and sets or resets based on the input.
+        snowflake name, their member object (which can be obtained from
+        `guild.fetch_member()`), and sets or resets based on the input.
 
         This will skip Peter, since Discord bots are not granted any
         permissions to change Server Owner nicknames, and will throw an
         error if they try to do so.
         """
-        if snowflake["name"] == "Peter": # prevent errors :-(
+        if snowflake["name"] == "Peter":  # prevent errors :-(
             return
 
         if reset:
             await member.edit(nick=member.name)
+            self.logger.info(f"Reset snowflake {snowflake['name']}'s nickname!")
         else:
             await member.edit(nick=snowflake["nickname"])
+            self.logger.info(
+                f"Set snowflake {snowflake['name']}'s nickname to {snowflake['nickname']}!"
+            )
 
     async def set_snowflake_names(self, guild, reset=False):
         """
         Sets or resets the nicknames for all snowflakes that have not
         retreated to their safe space. Requires the guild object passed in
-        either from ctx.message.guild or message.guild from the listener
+        either from `ctx.message.guild` or `message.guild` from the listener
         or command.
 
         The command will iterate through the snowflake list, setting or
@@ -60,6 +64,13 @@ class Snowflake(commands.Cog):
 
     @commands.Cog.listener()
     async def on_message(self, message):
+        """
+        The main listener routine that Snowflake Mode uses to respond to users.
+        The routine will check if Snowflake Mode is enabled, and send a message
+        if the user hasn't retreated to their safe space, as well as change their
+        nickname. If they type the `i made a doody` safe space message, it will
+        disable their Snowflake Mode and reset their nickname. 
+        """
         if message.author == self.bot.user:
             return
 
@@ -67,7 +78,7 @@ class Snowflake(commands.Cog):
             return
 
         if self.snowflake_mode:
-            if message.author.id in self.thelist.keys():
+            if message.author.id in self.thelist:
                 snowflake = self.thelist[message.author.id]
                 if self.snowflake_list[message.author.id]:
                     sleep(1)
@@ -81,45 +92,106 @@ class Snowflake(commands.Cog):
                             await message.channel.send(file=picture)
                         member = await message.guild.fetch_member(message.author.id)
                         await self.set_snowflake_name(snowflake, member, reset=True)
+                        self.logger.info(
+                            f"{snowflake.name} retreated to their safe space!"
+                        )
                     else:
                         response = f"{snowflake['message']}\n{snowflake['video']}"
                         await message.channel.send(response)
                         member = await message.guild.fetch_member(message.author.id)
                         await self.set_snowflake_name(snowflake, member)
+                        self.logger.info(
+                            f"Bugged {snowflake.name} in {message.channel.name}!"
+                        )
+
+    @commands.Cog.listener()
+    async def on_member_update(self, before, after):
+        """
+        Listen for members trying to update their nickname. Prevent all snowflakes from
+        updating their nickname unless Snowflake Mode is disabled or they have retreated
+        to their safe space.
+
+        Does not work for Peter, as bots cannot update nicknames for Server Owners.
+        """
+        if before.nickname == after.nickname:
+            return
+
+        if self.snowflake_mode and after.id in self.thelist:
+            # Does not work with Server Owners, unfortunately
+            if self.thelist[after.id]["name"] == "Peter":
+                return
+
+            if self.snowflake_list[after.id]:
+                self.set_snowflake_name(after.id, after)
 
     @commands.command()
     async def snowflake(self, ctx, arg="on"):
+        """
+        The main !snowflake command, that has three options:
+
+        `on` (default) - enable Snowflake Mode for all users that have not retreated to
+        their safe space, and change their nicknames.
+
+        `off` - disable Snowflake Mode for all users and reset all nicknames.
+
+        `force` - enable Snowflake Mode for all users, even those that have explicitly
+        retreated to their safe space. Can only be invoked by snowflakes!
+        """
         if arg == "on":
             self.snowflake_mode = True
             response = "Ooh, wee! We're gonna get some people pissed off, tonight!"
-            await self.set_snowflake_names(ctx.message.guild)
             await ctx.channel.send(response)
+            await self.set_snowflake_names(ctx.message.guild)
             with open(os.path.join(RESOURCES_DIR, "snowflake.jpg"), "rb") as file:
                 picture = discord.File(file)
                 await ctx.channel.send(file=picture)
+            self.logger.info(
+                f"{ctx.author.name} enabled Snowflake Mode in {ctx.channel.name}!"
+            )
 
         elif arg == "off":
             self.snowflake_mode = False
             response = "Ooh, wee! Looks like *all* the snowflakes need a break!"
-            await self.set_snowflake_names(ctx.message.guild, reset=True)
             await ctx.channel.send(response)
+            await self.set_snowflake_names(ctx.message.guild, reset=True)
             with open(os.path.join(RESOURCES_DIR, "privilege.jpg"), "rb") as file:
                 picture = discord.File(file)
                 await ctx.channel.send(file=picture)
+            self.logger.info(
+                f"{ctx.author.name} disabled Snowflake Mode in {ctx.channel.name}!"
+            )
 
         elif arg == "force":
+            if ctx.author.id not in self.thelist:
+                response = (
+                    "Only fellow snowflakes can force Snowflake Mode on! Ooh, wee!"
+                )
+                await ctx.channel.send(response)
+                self.logger.info(
+                    f"{ctx.author.name} tried to force on Snowflake Mode in {ctx.channel.name}!"
+                )
+                return
+
             self.snowflake_mode = True
             for snowflake in self.snowflake_list.keys():
                 self.snowflake_list[snowflake] = True
+
             response = "Ooh, wee! Time for all your safe spaces to burn down!"
-            await self.set_snowflake_names(ctx.message.guild, reset=True)
             await ctx.channel.send(response)
+            await self.set_snowflake_names(ctx.message.guild)
             with open(os.path.join(RESOURCES_DIR, "safespace.jpg"), "rb") as file:
                 picture = discord.File(file)
                 await ctx.channel.send(file=picture)
+            self.logger.info(
+                f"{ctx.author.name} disabled Snowflake Mode in {ctx.channel.name}!"
+            )
 
     @commands.command()
     async def snowflakes(self, ctx):
+        """
+        The !snowflakes command, which will print out some status info on Snowflake Mode,
+        such as whether it is enabled, and what users have retreated to their safe space.
+        """
         response = "Ooh, wee! Let's see how the snowflakes are doing!"
         await ctx.channel.send(response)
         with open(os.path.join(RESOURCES_DIR, "snowflakes.jpg"), "rb") as file: